% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{distri_data}
\alias{distri_data}
\title{Additions}
\usage{
distri_data(base, x, y, groupe)
}
\arguments{
\item{base}{nom de la base de donnees sur laquelle}

\item{x}{variable x}

\item{y}{variable y}

\item{groupe}{variable optionnelle si on veut grouper en fonction d'un facteur}
}
\value{
renvoie un boxplot des données
}
\description{
Longue description du package
}
\examples{

}
