% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fn_r}
\alias{fn_r}
\title{Compute intrinsic growth rates for a given model}
\usage{
fn_r(model)
}
\arguments{
\item{model}{Model type: "glv", "ricker", or "bh".}
}
\value{
A function that computes r given alpha and x0.
}
\description{
Compute intrinsic growth rates for a given model
}
