% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gestion_repertoire.R
\name{import_data}
\alias{import_data}
\title{Additions}
\usage{
import_data(rep, file)
}
\arguments{
\item{rep}{nom du repetoire ou se trouvent le donnees}

\item{file}{nom du fichier avec extension .Rdata}
}
\value{
importe le fichier rdata
}
\description{
Longue description du package
}
\examples{

}
